+++ Plugin/Plugin.vcxproj
@@ -106,7 +106,7 @@
     <ClCompile Include="mapAdj.cpp" />
     <ClCompile Include="NickNameFix.cpp" />
     <ClCompile Include="NoDynastyId.cpp" />
-    <ClCompile Include="Unk1.cpp" />
+    <ClCompile Include="MainTextLineBreak.cpp" />
     <ClCompile Include="MapAdj3.cpp" />
     <ClCompile Include="Unk3.cpp" />
     <ClCompile Include="Unk4.cpp" />
